Raggruppate query, aggiunto update della lista piani sulla creazione del piano
